+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from "./classes/rolling";
+export * from "./classes/rollUtils"
 export * from "./entities/dice";