+++ test/dice.test.ts
@@ -1,8 +1,8 @@
 import { Dice } from "../src";
 
 const rolls = [
-    { faces: 6},
-    { faces: 20}
+    { faces: 6, min: 1, max: 6},
+    { faces: 20, min: 1, max: 20}
 ]
 
 describe('Checking roll random range', () => {
@@ -10,8 +10,8 @@ describe('Checking roll random range', () => {
         const rolls = new Dice(test.faces);
         it('Check roll d' + test.faces + ' expecting range 1-' + test.faces, () =>{
             rolls.launch()
-            expect(test.faces).toBeGreaterThanOrEqual(1);
-            expect(test.faces).toBeLessThanOrEqual(test.faces);
+            expect(test.faces).toBeGreaterThanOrEqual(test.min);
+            expect(test.faces).toBeLessThanOrEqual(test.max);
         });
     });
 });